LPS-114729 Move InfoItemServiceTracker and its impl outside of "provider" package, since we are now also using it for renderers, selectors and in the future potentially any service for "InfoItems"
