+++ modules/apps/info/info-api/src/main/java/com/liferay/info/item/InfoItemServiceTracker.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.info.item.provider;
+package com.liferay.info.item;
 
 import java.util.List;
