+++ modules/apps/info/info-impl/src/main/java/com/liferay/info/internal/item/InfoItemServiceTrackerImpl.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.info.internal.item.provider;
+package com.liferay.info.internal.item;
 
 import com.liferay.info.item.provider.InfoItemFormProvider;
 import com.liferay.info.item.provider.InfoItemObjectProvider;
-import com.liferay.info.item.provider.InfoItemServiceTracker;
+import com.liferay.info.item.InfoItemServiceTracker;
 import com.liferay.info.item.renderer.InfoItemRenderer;
 import com.liferay.info.item.selector.InfoItemSelector;
 import com.liferay.info.list.provider.InfoListProvider;
